requirements.txt
flask==2.3.3
pandas==2.1.1
numpy==1.25.2
scikit-learn==1.3.1
pdfplumber==0.10.3
PyPDF2==3.0.1
matplotlib==3.8.0
shap==0.43.0
joblib==1.3.2
Werkzeug==2.3.7
